vivid: Switch to jellybeans from gruvbox-dark
Normal text with a background[] bleeds into the rest of the terminal when
used in zsh list-colors. This isn't a problem when the terminal theme
matches. I switch between the dark and light variants for my
terminal based on ambient light, and keeping LS_COLORS in sync does not
seem feasible.

[]: https://github.com/sharkdp/vivid/commit/dc0cfc7152ede98144ab98f4b5004cb9e36f4356#diff-93f729c80233c2b01f90841aae6eb84ca47eb71bb72a188825ad664ca839b060R37